Update window_panel.py
The current window panel is alright for a one-time, static window panel. But if you want to have it regularly updated, then I guess you need to destroy the existing window panel and create a new one at certain intervals. The problem is with the dragging: if you drag a window panel to a certain location, the update will always bring it back to the center. This is annoying especially if you have multiple window panels. The proposed changes allows one to easily copy the position (position.xy) of the old window panel to the new updated window panel.