Added python sqrt with Fast Inverse Square Root
The fast inverse Square root algorithm is one of the fastest if not the fastest way to do square root. This incredibly speeds up the FPS and optimizes the game engine. Game engines like Unity, Unreal and even pygame use the Fast Inverse Square Root Algorithm. It fast because in the way it uses pointers effectively. That is why we need to use ctypes which is part of the Python standard library so it is very compatitable with Python and will minimize the chance of a device not being supported when export through modules like pyinstaller & py2exe